Packagify operations too
